g3bench: use varint correctly in thrift theader

+++ lib/g3-types/src/codec/thrift/var_int.rs
@@ -17,10 +17,13 @@ impl ThriftVarInt32 {
         Ok(ThriftVarInt32 { leb128 })
     }
 
-    pub fn positive_value(&self) -> u32 {
-        self.leb128.value()
+    // Get the varint value used directly in thrift protocol,
+    // which is always positive and not zigzag encoded
+    pub fn positive_value(&self) -> i32 {
+        self.leb128.value() as i32
     }
 
+    // Get the thrift integer value, which is zigzag encoded
     pub fn value(&self) -> i32 {
         let uv = self.leb128.value();
         i32::from_zig_zag(uv)
@@ -37,12 +40,14 @@ pub struct ThriftVarIntEncoder {
 }
 
 impl ThriftVarIntEncoder {
+    // Encode the thrift integer value, with correct zigzag encoding
     pub fn encode_i32(&mut self, v: i32) -> &[u8] {
         let uv = v.to_zig_zag();
         self.leb128.encode_u32(uv)
     }
 
-    pub fn encode_positive_i32(&mut self, v: u32) -> &[u8] {
-        self.leb128.encode_u32(v)
+    // Encode the positive varint used directly in thrift protocol, which will not be zigzag encoded
+    pub fn encode_positive_i32(&mut self, v: i32) -> &[u8] {
+        self.leb128.encode_u32(v as u32)
     }
 }
